+++ resources/js/components/common/DeleteModal.vue
@@ -7,23 +7,20 @@
 					<div class="close" data-dismiss="modal">&times;</div>
 				</div>
 				
-				<div class="modal-body">
+				<div class="modal-body" style="margin-bottom:10px">
 					<form method="POST" @submit.prevent="deleteRecord">
 						<input type="hidden" name="_method" value="delete" />
 						<div class="emp-label">
-							<center>Are you sure you want to remove this record?</center>
+							<center>Are you sure you want to delete this record?</center>
 						</div>
 					</form>
 				</div>
 
 				<div class="modal-footer" style="border-top:none">
-					<div class="btn btn-link btn-delete" data-dismiss="modal">
-						Cancel
+					<div class="pull-right">
+						<button type="submit" @click="deleteRecord" :disabled="disabled">Yes</button>
+						<button @click="cancel">No</button>
 					</div>
-
-					<button class="pull-right" type="submit" @click="deleteRecord" :disabled="disabled">
-						Proceed
-					</button>
 				</div>
 			</div>
         </div>
@@ -52,20 +49,30 @@
 			initForm() {
 				let component = this;
 
+				Bus.$on('deleteAboutMe', function() {
+					component.action = 'AboutMe';
+					component.endpoints.delete = '';
+				});
+
+				Bus.$on('deleteIdealRole', function() {
+					component.action = 'IdealRole';
+					component.endpoints.delete = '';
+				});
+
 				Bus.$on('deleteEmployment', function(index, endpoint) {
-					component.action = 'deleteEmployment';
+					component.action = 'Employment';
 					component.index = index;
 					component.endpoints.delete = endpoint;
 				});
 
 				Bus.$on('deleteEducation', function(index, endpoint) {
-					component.action = 'deleteEducation';
+					component.action = 'Education';
 					component.index = index;
 					component.endpoints.delete = endpoint;
 				});
 
 				Bus.$on('deleteIndustrySkill', function(endpoint) {
-					component.action = 'deleteIndustrySkill';
+					component.action = 'IndustrySkill';
 					component.endpoints.delete = endpoint;
 				});
 			},
@@ -75,32 +82,38 @@
 
 				this.disabled = true;
 				
-				await axios.delete(component.endpoints.delete, Utils.getBearerAuth())
+				if (this.action == 'Employment' || this.action == 'Education' || this.action == 'IndustrySkill') {
 
-					.then(function(response) {
-						let data = response.data;
-							
-						if (data.success) {
-							$('#deleteRecordModal').modal('hide');
-							console.log(component.index)
+					await axios.delete(component.endpoints.delete, Utils.getBearerAuth())
 
-							if (component.action == 'deleteEmployment') {
-								Bus.$emit('removeEmployment', component.index);
-							
-							} else if (component.action == 'deleteEducation') {
-								Bus.$emit('removeEducation', component.index);	
-							
-							} else if(component.action == 'deleteIndustrySkill') {
-								Bus.$emit('removeIndustrySkill');
+						.then(function(response) {
+							let data = response.data;
+								
+							if (data.success) {
+								$('#deleteRecordModal').modal('hide');
+								
+								Bus.$emit('remove' + component.action, component.index);
 							}
-						}
-					}).catch(function(error) {
-						
-						Utils.handleError(error);
-					});
+						}).catch(function(error) {
+							
+							Utils.handleError(error);
+						});
+				
+				} else {
+					$('#deleteRecordModal').modal('hide');
+
+					Bus.$emit('remove' + component.action);
+				}
 				
 				this.disabled = false;
 			},
+
+			cancel() {
+				$('#deleteRecordModal').modal('hide');
+
+				$('#modal' + this.action).modal('show');
+			},
+
 		}
 	}
 </script>
